refactor(api): split server.js into modules (Stage I)
Decomposes the 2910-line `backend/src/api/server.js` into focused
modules. Server.js becomes a 307-line bootstrap that wires services,
middleware, GraphQL, and routers together.

New modules:
- schema/sdl.js          — GraphQL SDL + buildSchema
- resolvers/index.js     — createResolvers({db, store})
- middleware/index.js    — installBaseMiddleware, buildWriteRateLimiter,
                           buildRequireApiKey, installGraphQL,
                           installErrorHandling
- routes/health.js       — /health, /api/status, /api/stats
- routes/events.js       — /api/events/{prepare,create,store-for-anchor,
                           confirm-anchor,:hash,dev-pubkey,dev-sign}
- routes/ingest.js       — /api/ingest/anchored-event
- routes/crypto.js       — /api/crypto/resolve-signing-key
- routes/entities.js     — /api/{person,group,track,release,song,label,
                           groups}/...
- routes/search.js       — /api/search/nodes
- routes/curate.js       — /api/curate/operations* + parseOperationDetail
                           helper (was a private method on APIServer)
- routes/player.js       — /api/player/queue
- routes/graph.js        — /api/graph/{initial,neighborhood/:nodeId}

Each route factory accepts a context object ({db, store, config,
ingestionHandler, playerService, writeRateLimiter, requireApiKey})
rather than reaching through `this`, so the modules are testable in
isolation.

Behaviour-preserving: zero changes to URL paths, response shapes,
status codes, or middleware ordering. Verified via the Stage H
characterization tests (484 passed / 218 skipped / 30 snapshots
matched — identical to baseline).
