+++ backend/src/api/server.js
@@ -1,272 +1,74 @@
 /**
- * @fileoverview Main API server for Polaris Music Registry
+ * @fileoverview Main API server for Polaris Music Registry.
  *
- * Provides both GraphQL and REST endpoints for:
- * - Querying music data from the graph database
- * - Submitting and retrieving events
- * - Getting statistics and metadata
+ * Bootstraps the Express app and wires together:
+ *   - Domain services (graph database, event store, indexer/ingestion,
+ *     player, chain reader)
+ *   - Middleware (security headers, CORS, JSON, request logging,
+ *     write rate limiter, ingestion API-key)
+ *   - GraphQL endpoint at /graphql
+ *   - REST endpoints under /health, /api/...
  *
- * Architecture:
- * - GraphQL for flexible queries (groups, persons, releases, tracks)
- * - REST for specific operations (event submission, stats)
- * - Integrates with MusicGraphDatabase and EventStore
+ * The route handlers and resolvers themselves live in sibling modules
+ * (see ./schema, ./resolvers, ./middleware, ./routes/*) — this file is
+ * intentionally thin so adding a new endpoint never requires editing a
+ * 3000-line file again.
  *
  * @module api/server
  */
 
 import express from 'express';
-import cors from 'cors';
-import helmet from 'helmet';
-import rateLimit from 'express-rate-limit';
-import { graphqlHTTP } from 'express-graphql';
-import { buildSchema } from 'graphql';
 import MusicGraphDatabase from '../graph/schema.js';
 import EventStore from '../storage/eventStore.js';
 import EventProcessor from '../indexer/eventProcessor.js';
-import { createIdentityRoutes } from './routes/identity.js';
-import { normalizeReleaseBundle } from '../graph/normalizeReleaseBundle.js';
-import { validateReleaseBundleOrThrow } from '../schema/validateReleaseBundle.js';
-
 import { IngestionHandler } from './ingestion.js';
 import { PlayerService } from './playerService.js';
-import { NodeSearchService } from './nodeSearchService.js';
 import { ChainReaderService } from './chainReaderService.js';
 import { getDevSigner } from '../crypto/devSigner.js';
-import { getStatus } from './status.js';
-import { createHash, timingSafeEqual } from 'crypto';
-import { PublicKey, Signature } from 'eosjs/dist/eosjs-key-conversions.js';
-import { createLogger, generateRequestId } from '../utils/logger.js';
-import { sanitizeError } from '../utils/errorSanitizer.js';
-import neo4j from 'neo4j-driver';
+
+import { schema } from './schema/sdl.js';
+import { createResolvers } from './resolvers/index.js';
+import {
+    installBaseMiddleware,
+    buildWriteRateLimiter,
+    buildRequireApiKey,
+    installGraphQL,
+    installErrorHandling,
+} from './middleware/index.js';
+
+import { createIdentityRoutes } from './routes/identity.js';
+import { createHealthRoutes } from './routes/health.js';
+import { createEventsRoutes } from './routes/events.js';
+import { createIngestRoutes } from './routes/ingest.js';
+import { createCryptoRoutes } from './routes/crypto.js';
+import { createEntityRoutes } from './routes/entities.js';
+import { createSearchRoutes } from './routes/search.js';
+import { createCurateRoutes } from './routes/curate.js';
+import { createPlayerRoutes } from './routes/player.js';
+import { createGraphRoutes } from './routes/graph.js';
 
 /**
- * GraphQL Schema Definition
- * Defines all types, queries, and mutations available via GraphQL
- */
-const schema = buildSchema(`
-  """
-  A person in the music industry (musician, producer, engineer, etc.)
-  """
-  type Person {
-    person_id: String!
-    name: String!
-    alt_names: [String!]
-    bio: String
-    status: String!
-    groups: [GroupMembership!]
-    songsWritten: [Song!]
-    tracksProduced: [Track!]
-    guestAppearances: [Track!]
-  }
-
-  """
-  Group membership details with dates and roles
-  """
-  type GroupMembership {
-    group: Group!
-    role: String
-    from_date: String
-    to_date: String
-    instruments: [String!]
-  }
-
-  """
-  A musical group (band, orchestra, ensemble, or solo project)
-  """
-  type Group {
-    group_id: String!
-    name: String!
-    alt_names: [String!]
-    bio: String
-    formed_date: String
-    disbanded_date: String
-    inferred_first_release_date: String
-    inferred_last_release_date: String
-    status: String!
-    members: [Member!]
-    releases: [Release!]
-    tracks: [Track!]
-  }
-
-  """
-  Member participation data for RGraph visualization
-  """
-  type Member {
-    person: Person!
-    role: String
-    from_date: String
-    to_date: String
-    instruments: [String!]
-    participation_percentage: Float
-    track_count: Int
-    release_count: Int
-  }
-
-  """
-  A musical composition (the written work)
-  """
-  type Song {
-    song_id: String!
-    title: String!
-    alt_titles: [String!]
-    iswc: String
-    year: Int
-    lyrics: String
-    writers: [Person!]
-    recordings: [Track!]
-  }
-
-  """
-  A specific recording/performance of a song
-  """
-  type Track {
-    track_id: String!
-    title: String!
-    isrc: String
-    duration: Int
-    recording_date: String
-    recording_location: String
-    listen_links: [String!]
-    status: String!
-    performedBy: [Group!]
-    guests: [Person!]
-    recordingOf: Song
-    releases: [Release!]
-  }
-
-  """
-  An album, EP, single, or other release package
-  """
-  type Release {
-    release_id: String!
-    name: String!
-    alt_names: [String!]
-    release_date: String
-    format: [String!]
-    country: String
-    catalog_number: String
-    liner_notes: String
-    album_art: String
-    status: String!
-    tracks: [TrackInRelease!]
-    labels: [Label!]
-    master: Master
-  }
-
-  """
-  Track position in a release with ordering
-  """
-  type TrackInRelease {
-    track: Track!
-    disc_number: Int
-    track_number: Int!
-    side: String
-    is_bonus: Boolean
-  }
-
-  """
-  Record label
-  """
-  type Label {
-    label_id: String!
-    name: String!
-    alt_names: [String!]
-    status: String!
-  }
-
-  """
-  Canonical album grouping for re-releases
-  """
-  type Master {
-    master_id: String!
-    name: String!
-  }
-
-  """
-  Search result union type
-  """
-  union SearchResult = Person | Group | Release | Track | Song
-
-  """
-  Statistics about the database
-  """
-  type Stats {
-    nodes: NodeStats!
-    enabled_services: EnabledServices!
-  }
-
-  type NodeStats {
-    Person: Int
-    Group: Int
-    Track: Int
-    Song: Int
-    Release: Int
-    Label: Int
-    total: Int
-  }
-
-  type EnabledServices {
-    ipfs: Boolean!
-    s3: Boolean!
-    redis: Boolean!
-  }
-
-  """
-  Root Query type
-  """
-  type Query {
-    """Get person by ID"""
-    person(person_id: String!): Person
-
-    """Get group by ID"""
-    group(group_id: String!): Group
-
-    """Get group member participation data (for RGraph)"""
-    groupParticipation(group_id: String!): [Member!]
-
-    """Get release by ID"""
-    release(release_id: String!): Release
-
-    """Get track by ID"""
-    track(track_id: String!): Track
-
-    """Get song by ID"""
-    song(song_id: String!): Song
-
-    """Search across all entity types"""
-    search(query: String!, limit: Int): [SearchResult!]
-
-    """Get database statistics"""
-    stats: Stats!
-
-    """Test database connectivity"""
-    testConnectivity: Boolean!
-  }
-
-`);
-
-/**
- * API Server class that manages Express app with GraphQL and REST endpoints
+ * API Server class that manages an Express app with GraphQL and REST endpoints.
  */
 class APIServer {
     /**
-     * Create a new API server
-     *
-     * @param {Object} config - Server configuration
-     * @param {number} config.port - Port to listen on
+     * @param {Object} config
+     * @param {number} [config.port=3000]
      * @param {Object} config.database - Graph database config
-     * @param {Object} config.storage - Event storage config
+     * @param {Object} config.storage  - Event storage config
+     * @param {string} [config.corsOrigin]
+     * @param {string} [config.rpcUrl]
+     * @param {string} [config.contractAccount]
+     * @param {string} [config.env]
      */
     constructor(config) {
         this.config = config;
         this.app = express();
         this.port = config.port || 3000;
 
-        // Initialize database and storage
+        // Domain services
         this.db = new MusicGraphDatabase(config.database);
         this.store = new EventStore(config.storage);
-
-        // Initialize event processor and ingestion handler (T5)
         this.eventProcessor = new EventProcessor({
             db: this.db,
             store: this.store
@@ -280,2380 +82,89 @@ class APIServer {
             contractAccount: process.env.CONTRACT_ACCOUNT || config.contractAccount || 'polarismusic'
         });
 
-        // Setup middleware and routes
-        this.setupMiddleware();
-        this.setupGraphQL();
-        this.setupRESTEndpoints();
-        this.setupErrorHandling();
+        // Wire middleware, GraphQL, and routes synchronously so that
+        // `new APIServer(...)` is enough to make `this.app` testable
+        // without awaiting start().
+        this._setup();
     }
 
-    /**
-     * Setup Express middleware
-     */
-    setupMiddleware() {
-        // Helmet - Security headers (XSS, CSP, etc.)
-        // Configure for compatibility with GraphiQL in development
-        this.app.use(helmet({
-            contentSecurityPolicy: process.env.NODE_ENV === 'production' ? undefined : false,
-            crossOriginEmbedderPolicy: false
-        }));
-        console.log(' Helmet security headers enabled');
+    _setup() {
+        // 1. Base middleware (helmet, cors, json, request-id logger)
+        installBaseMiddleware(this.app, this.config);
 
-        // CORS - Enable cross-origin requests from frontend
-        // Supports comma-separated list of origins for multiple dev environments
-        // Example: CORS_ORIGIN=http://localhost:5173,http://localhost:4173
-        const corsOriginEnv = this.config.corsOrigin || process.env.CORS_ORIGIN || 'http://localhost:5173';
-        const origins = corsOriginEnv
-            .split(',')
-            .map(s => s.trim())
-            .filter(Boolean);
+        // 2. Shared per-request guards used by multiple routers
+        this.writeRateLimiter = buildWriteRateLimiter();
+        this.requireApiKey = buildRequireApiKey();
 
-        // Pass single string or array to cors() depending on count
-        const corsOrigin = origins.length === 1 ? origins[0] : origins;
+        // 3. GraphQL endpoint
+        const rootValue = createResolvers({ db: this.db, store: this.store });
+        installGraphQL(this.app, schema, rootValue);
 
-        this.app.use(cors({
-            origin: corsOrigin,
-            methods: ['GET', 'POST', 'PUT', 'DELETE', 'OPTIONS'],
-            allowedHeaders: ['Origin', 'X-Requested-With', 'Content-Type', 'Accept', 'Authorization', 'X-API-Key'],
-            credentials: false
-        }));
-        console.log(` CORS enabled for origin(s): ${origins.join(', ')}`);
+        // 4. REST endpoints
+        this._installRESTRoutes();
 
-        // Parse JSON bodies
-        this.app.use(express.json({ limit: '10mb' }));
-
-        // Structured request logging with correlation ID
-        this.app.use((req, res, next) => {
-            const requestId = req.headers['x-request-id'] || generateRequestId();
-            req.requestId = requestId;
-            res.setHeader('X-Request-Id', requestId);
-
-            const log = createLogger('api.server', { request_id: requestId });
-            const timer = log.startTimer();
-
-            log.info('request_start', {
-                method: req.method,
-                path: req.path,
-                remote_ip: req.ip,
-                user_agent: req.get('user-agent')
-            });
-
-            const originalEnd = res.end;
-            res.end = function(...args) {
-                timer.end('request_end', {
-                    method: req.method,
-                    path: req.path,
-                    status: res.statusCode
-                });
-                originalEnd.apply(res, args);
-            };
-
-            next();
-        });
-
-        // Rate limiting for write routes (coarse, IP-based)
-        // Prevents storage cost blow-up and graph pollution from spammy clients
-        this.writeRateLimiter = rateLimit({
-            windowMs: 60 * 1000, // 1-minute window
-            max: 30,             // 30 write requests per minute per IP
-            standardHeaders: true,
-            legacyHeaders: false,
-            message: { error: 'Too many write requests, please try again later' }
-        });
-
-        // API key middleware for ingestion endpoints only.
-        // Applied to: /api/ingest/anchored-event
-        // NOT applied to: /api/events/prepare, /api/events/create (public, signature-verified)
-        // When INGEST_API_KEY is set, protected endpoints require X-API-Key header.
-        // When unset (local dev), protected endpoints are open.
-        this.requireApiKey = (req, res, next) => {
-            const requiredKey = process.env.INGEST_API_KEY;
-            if (!requiredKey) {
-                // In chain mode / production, missing key is a server misconfiguration
-                const ingestMode = process.env.INGEST_MODE || 'chain';
-                if (ingestMode === 'chain' || process.env.NODE_ENV === 'production') {
-                    return res.status(500).json({
-                        error: 'Server misconfigured: INGEST_API_KEY is required in chain/production mode'
-                    });
-                }
-                return next(); // No key configured — open access (dev mode only)
-            }
-            const provided = req.headers['x-api-key'];
-            if (!provided) {
-                return res.status(401).json({ error: 'Missing X-API-Key header' });
-            }
-            // Constant-time comparison to prevent timing attacks
-            if (provided.length !== requiredKey.length ||
-                !timingSafeEqual(
-                    Buffer.from(provided),
-                    Buffer.from(requiredKey)
-                )) {
-                return res.status(403).json({ error: 'Invalid API key' });
-            }
-            next();
-        };
+        // 5. 404 + global error handlers (must be last)
+        installErrorHandling(this.app);
     }
 
-    /**
-     * Setup GraphQL endpoint with resolvers
-     */
-    setupGraphQL() {
-        const root = {
-            // ========== PERSON QUERIES ==========
-            person: async ({ person_id }) => {
-                const session = this.db.driver.session();
-                try {
-                    const result = await session.run(`
-                        MATCH (p:Person {person_id: $id})
-                        WHERE p.status = 'ACTIVE'
-                        OPTIONAL MATCH (p)-[m:MEMBER_OF]->(g:Group)
-                        WHERE g.status = 'ACTIVE'
-                        OPTIONAL MATCH (p)-[:WROTE]->(s:Song)
-                        WHERE s.status = 'ACTIVE'
-                        OPTIONAL MATCH (p)-[:PRODUCED]->(t:Track)
-                        WHERE t.status = 'ACTIVE'
-                        OPTIONAL MATCH (p)-[:GUEST_ON]->(tg:Track)
-                        WHERE tg.status = 'ACTIVE'
-
-                        RETURN p,
-                               collect(DISTINCT {
-                                   group: g,
-                                   role: m.role,
-                                   from_date: m.from_date,
-                                   to_date: m.to_date,
-                                   instruments: m.instruments
-                               }) as groups,
-                               collect(DISTINCT s) as songsWritten,
-                               collect(DISTINCT t) as tracksProduced,
-                               collect(DISTINCT tg) as guestAppearances
-                    `, { id: person_id });
-
-                    if (result.records.length === 0) return null;
-
-                    const record = result.records[0];
-                    const person = record.get('p').properties;
-                    const groups = record.get('groups')
-                        .filter(g => g.group !== null)
-                        .map(g => ({
-                            group: g.group.properties,
-                            role: g.role,
-                            from_date: g.from_date,
-                            to_date: g.to_date,
-                            instruments: g.instruments || []
-                        }));
-                    const songsWritten = record.get('songsWritten')
-                        .filter(s => s !== null)
-                        .map(s => s.properties);
-                    const tracksProduced = record.get('tracksProduced')
-                        .filter(t => t !== null)
-                        .map(t => t.properties);
-                    const guestAppearances = record.get('guestAppearances')
-                        .filter(t => t !== null)
-                        .map(t => t.properties);
-
-                    return {
-                        ...person,
-                        alt_names: person.alt_names || [],
-                        groups,
-                        songsWritten,
-                        tracksProduced,
-                        guestAppearances
-                    };
-                } finally {
-                    await session.close();
-                }
-            },
-
-            // ========== GROUP QUERIES ==========
-            group: async ({ group_id }) => {
-                const session = this.db.driver.session();
-                try {
-                    const result = await session.run(`
-                        MATCH (g:Group {group_id: $id})
-                        WHERE g.status = 'ACTIVE'
-                        OPTIONAL MATCH (p:Person)-[m:MEMBER_OF]->(g)
-                        WHERE p.status = 'ACTIVE'
-                        OPTIONAL MATCH (g)-[:PERFORMED_ON]->(t:Track)
-                        WHERE t.status = 'ACTIVE'
-                        OPTIONAL MATCH (t)-[:IN_RELEASE]->(r:Release)
-                        WHERE r.status = 'ACTIVE'
-
-                        RETURN g,
-                               collect(DISTINCT {
-                                   person: p,
-                                   role: m.role,
-                                   from_date: m.from_date,
-                                   to_date: m.to_date,
-                                   instruments: m.instruments
-                               }) as members,
-                               collect(DISTINCT t) as tracks,
-                               collect(DISTINCT r) as releases
-                    `, { id: group_id });
-
-                    if (result.records.length === 0) return null;
-
-                    const record = result.records[0];
-                    const group = record.get('g').properties;
-                    const members = record.get('members')
-                        .filter(m => m.person !== null)
-                        .map(m => ({
-                            person: m.person.properties,
-                            role: m.role,
-                            from_date: m.from_date,
-                            to_date: m.to_date,
-                            instruments: m.instruments || []
-                        }));
-                    const tracks = record.get('tracks')
-                        .filter(t => t !== null)
-                        .map(t => t.properties);
-                    const releases = record.get('releases')
-                        .filter(r => r !== null)
-                        .map(r => r.properties);
-
-                    // Compute inferred active dates from release dates
-                    const releaseDates = releases
-                        .map(r => r.release_date)
-                        .filter(d => d != null && d !== '');
-                    const inferred_first_release_date = releaseDates.length > 0
-                        ? releaseDates.reduce((a, b) => a < b ? a : b)
-                        : null;
-                    const inferred_last_release_date = releaseDates.length > 0
-                        ? releaseDates.reduce((a, b) => a > b ? a : b)
-                        : null;
-
-                    return {
-                        ...group,
-                        alt_names: group.alt_names || [],
-                        inferred_first_release_date,
-                        inferred_last_release_date,
-                        members,
-                        tracks,
-                        releases
-                    };
-                } finally {
-                    await session.close();
-                }
-            },
-
-            groupParticipation: async ({ group_id }) => {
-                const participation = await this.db.calculateGroupMemberParticipation(group_id);
-                return participation.map(p => ({
-                    person: { person_id: p.personId, name: p.personName },
-                    participation_percentage: p.trackPctOfGroupTracks,
-                    track_count: p.trackCount,
-                    release_count: null
-                }));
-            },
-
-            // ========== RELEASE QUERIES ==========
-            release: async ({ release_id }) => {
-                const session = this.db.driver.session();
-                try {
-                    const result = await session.run(`
-                        MATCH (r:Release {release_id: $id})
-                        WHERE r.status = 'ACTIVE'
-                        OPTIONAL MATCH (t:Track)-[ir:IN_RELEASE]->(r)
-                        WHERE t.status = 'ACTIVE'
-                        OPTIONAL MATCH (r)-[:RELEASED]->(l:Label)
-                        WHERE l.status = 'ACTIVE'
-                        OPTIONAL MATCH (r)-[:IN_MASTER]->(m:Master)
-                        WHERE m.status = 'ACTIVE'
-
-                        RETURN r,
-                               collect(DISTINCT {
-                                   track: t,
-                                   disc_number: ir.disc_number,
-                                   track_number: ir.track_number,
-                                   side: ir.side,
-                                   is_bonus: ir.is_bonus
-                               }) as tracks,
-                               collect(DISTINCT l) as labels,
-                               m as master
-                    `, { id: release_id });
-
-                    if (result.records.length === 0) return null;
-
-                    const record = result.records[0];
-                    const release = record.get('r').properties;
-                    const tracks = record.get('tracks')
-                        .filter(t => t.track !== null)
-                        .map(t => ({
-                            track: t.track.properties,
-                            disc_number: t.disc_number,
-                            track_number: t.track_number,
-                            side: t.side,
-                            is_bonus: t.is_bonus
-                        }));
-                    const labels = record.get('labels')
-                        .filter(l => l !== null)
-                        .map(l => l.properties);
-                    const master = record.get('master');
-
-                    return {
-                        ...release,
-                        alt_names: release.alt_names || [],
-                        format: release.format || [],
-                        tracks,
-                        labels,
-                        master: master ? master.properties : null
-                    };
-                } finally {
-                    await session.close();
-                }
-            },
-
-            // ========== TRACK QUERIES ==========
-            track: async ({ track_id }) => {
-                const session = this.db.driver.session();
-                try {
-                    const result = await session.run(`
-                        MATCH (t:Track {track_id: $id})
-                        WHERE t.status = 'ACTIVE'
-                        OPTIONAL MATCH (t)-[:RECORDING_OF]->(s:Song)
-                        WHERE s.status = 'ACTIVE'
-                        OPTIONAL MATCH (g:Group)-[:PERFORMED_ON]->(t)
-                        WHERE g.status = 'ACTIVE'
-                        OPTIONAL MATCH (p:Person)-[:GUEST_ON]->(t)
-                        WHERE p.status = 'ACTIVE'
-                        OPTIONAL MATCH (t)-[:IN_RELEASE]->(r:Release)
-                        WHERE r.status = 'ACTIVE'
-
-                        RETURN t,
-                               s,
-                               collect(DISTINCT g) as performedBy,
-                               collect(DISTINCT p) as guests,
-                               collect(DISTINCT r) as releases
-                    `, { id: track_id });
-
-                    if (result.records.length === 0) return null;
-
-                    const record = result.records[0];
-                    const track = record.get('t').properties;
-                    const song = record.get('s');
-                    const performedBy = record.get('performedBy')
-                        .filter(g => g !== null)
-                        .map(g => g.properties);
-                    const guests = record.get('guests')
-                        .filter(p => p !== null)
-                        .map(p => p.properties);
-                    const releases = record.get('releases')
-                        .filter(r => r !== null)
-                        .map(r => r.properties);
-
-                    return {
-                        ...track,
-                        listen_links: track.listen_links || [],
-                        recordingOf: song ? song.properties : null,
-                        performedBy,
-                        guests,
-                        releases
-                    };
-                } finally {
-                    await session.close();
-                }
-            },
-
-            // ========== SONG QUERIES ==========
-            song: async ({ song_id }) => {
-                const session = this.db.driver.session();
-                try {
-                    const result = await session.run(`
-                        MATCH (s:Song {song_id: $id})
-                        WHERE s.status = 'ACTIVE'
-                        OPTIONAL MATCH (p:Person)-[:WROTE]->(s)
-                        WHERE p.status = 'ACTIVE'
-                        OPTIONAL MATCH (t:Track)-[:RECORDING_OF]->(s)
-                        WHERE t.status = 'ACTIVE'
-
-                        RETURN s,
-                               collect(DISTINCT p) as writers,
-                               collect(DISTINCT t) as recordings
-                    `, { id: song_id });
-
-                    if (result.records.length === 0) return null;
-
-                    const record = result.records[0];
-                    const song = record.get('s').properties;
-                    const writers = record.get('writers')
-                        .filter(p => p !== null)
-                        .map(p => p.properties);
-                    const recordings = record.get('recordings')
-                        .filter(t => t !== null)
-                        .map(t => t.properties);
-
-                    return {
-                        ...song,
-                        alt_titles: song.alt_titles || [],
-                        writers,
-                        recordings
-                    };
-                } finally {
-                    await session.close();
-                }
-            },
-
-            // ========== SEARCH ==========
-            search: async ({ query, limit = 10 }) => {
-                const searchService = new NodeSearchService(this.db.driver);
-                const results = await searchService.search(query, { limit });
-                return results.map(r => ({
-                    __typename: r.type,
-                    id: r.id,
-                    name: r.display_name,
-                    type: r.type,
-                    ...r
-                }));
-            },
-
-            // ========== STATS ==========
-            stats: async () => {
-                const dbStats = await this.db.getStats();
-                const storageStats = this.store.getStats();
-
-                const total = Object.values(dbStats.nodes).reduce((sum, count) => sum + count, 0);
-
-                return {
-                    nodes: {
-                        ...dbStats.nodes,
-                        total
-                    },
-                    enabled_services: storageStats.enabled
-                };
-            },
-
-            testConnectivity: async () => {
-                const dbConnected = await this.db.testConnection();
-                return dbConnected;
-            }
+    _installRESTRoutes() {
+        const ctx = {
+            db: this.db,
+            store: this.store,
+            ingestionHandler: this.ingestionHandler,
+            playerService: this.playerService,
+            config: this.config,
+            writeRateLimiter: this.writeRateLimiter,
+            requireApiKey: this.requireApiKey,
         };
 
-        // Setup GraphQL endpoint with structured logging
-        this.app.use('/graphql', (req, res, next) => {
-            const gqlLog = createLogger('api.graphql', { request_id: req.requestId });
-            const timer = gqlLog.startTimer();
-
-            const originalEnd = res._origEnd || res.end;
-            res.end = function(...args) {
-                const operationName = req.body?.operationName || undefined;
-                timer.end('graphql_request', {
-                    operation_name: operationName,
-                    status: res.statusCode
-                });
-                originalEnd.apply(res, args);
-            };
-
-            graphqlHTTP({
-                schema: schema,
-                rootValue: root,
-                graphiql: process.env.NODE_ENV !== 'production',
-                customFormatErrorFn: (error) => {
-                    gqlLog.error('graphql_error', {
-                        message: error.message,
-                        path: error.path
-                    });
-                    return {
-                        message: error.message,
-                        locations: error.locations,
-                        path: error.path
-                    };
-                }
-            })(req, res, next);
-        });
-
-        console.log(' GraphQL endpoint configured at /graphql');
-    }
-
-    /**
-     * Setup REST endpoints for specific operations
-     */
-    setupRESTEndpoints() {
-        // ========== IDENTITY MANAGEMENT ==========
-        // Mount identity routes
-        const identityRouter = createIdentityRoutes(this.db, this.store, this.eventProcessor);
-        this.app.use('/api/identity', identityRouter);
+        // Identity management
+        this.app.use(
+            '/api/identity',
+            createIdentityRoutes(this.db, this.store, this.eventProcessor)
+        );
         console.log(' Identity management endpoints mounted at /api/identity');
 
-        // ========== HEALTH CHECK ==========
-        this.app.get('/health', (req, res) => {
-            res.json({
-                status: 'healthy',
-                timestamp: new Date().toISOString(),
-                uptime: process.uptime()
-            });
-        });
+        // Health, status, stats — three separate paths sharing a factory
+        const { healthRouter, statusRouter, statsRouter } = createHealthRoutes(ctx);
+        this.app.use('/health', healthRouter);
+        this.app.use('/api/status', statusRouter);
+        this.app.use('/api/stats', statsRouter);
 
-        /**
-         * GET /api/status
-         * Comprehensive system status check for smoke testing and monitoring
-         *
-         * Checks all critical pipeline services and returns detailed health information.
-         *
-         * Response format:
-         * {
-         *   "ok": boolean,              // true if all critical services are healthy
-         *   "timestamp": "ISO8601",
-         *   "services": {
-         *     "ipfs": [                 // Array of IPFS nodes with details
-         *       {
-         *         "url": "http://ipfs:5001",
-         *         "ok": true,
-         *         "id": "QmXxxx...",    // IPFS peer ID
-         *         "version": "0.24.0"    // IPFS version
-         *       }
-         *     ],
-         *     "neo4j": { "ok": true },
-         *     "redis": { "ok": true },
-         *     "pinning_provider": {
-         *       "enabled": false,
-         *       "provider": "none"
-         *     }
-         *   }
-         * }
-         *
-         * Returns:
-         * - 200 with ok:true if all critical services are healthy
-         * - 503 with ok:false if any critical service is down
-         */
-        this.app.get('/api/status', async (req, res) => {
-            try {
-                const status = await getStatus({
-                    eventStore: this.store,
-                    neo4jDriver: this.db?.driver,
-                    redisClient: this.store?.redis,
-                    pinningProvider: this.store?.pinningProvider
-                });
+        // Event preparation, creation, retrieval, dev-signing helpers
+        this.app.use('/api/events', createEventsRoutes(ctx));
 
-                // Return 503 if any service is down, 200 if all OK
-                const httpStatus = status.ok ? 200 : 503;
-                res.status(httpStatus).json(status);
-            } catch (error) {
-                console.error('Status check failed:', error);
-                res.status(500).json(sanitizeError(error, req.requestId, {
-                    env: this.config.env,
-                }));
-            }
-        });
+        // Crypto helpers (resolve signing key)
+        this.app.use('/api/crypto', createCryptoRoutes(ctx));
 
-        // ========== EVENT ENDPOINTS ==========
-
-        /**
-         * POST /api/events/prepare
-         * Prepare an event for signing/anchoring by normalizing and returning canonical hash
-         *
-         * This endpoint allows the frontend to get the canonical hash that will be used
-         * for storage BEFORE signing, ensuring the hash anchored on-chain matches
-         * the hash the backend will store.
-         *
-         * Flow:
-         * 1. Frontend builds event (without sig)
-         * 2. Calls /api/events/prepare to get canonical hash
-         * 3. Signs the returned hash
-         * 4. Adds sig to event
-         * 5. Calls /api/events/create to store
-         *
-         * IMPORTANT: The event passed to /prepare must include author_pubkey.
-         * If you prepare without author_pubkey and add it later, the canonical
-         * payload/hash will not match what the verifier hashes (which includes
-         * author_pubkey), causing signature verification to fail.
-         *
-         * @returns {Object} { success: true, hash, normalizedEvent }
-         */
-        this.app.post('/api/events/prepare', this.writeRateLimiter, async (req, res) => {
-            try {
-                const event = req.body;
-
-                // Clone event to avoid mutating the original
-                const preparedEvent = JSON.parse(JSON.stringify(event));
-
-                // Normalize CREATE_RELEASE_BUNDLE events
-                if (preparedEvent.type === 'CREATE_RELEASE_BUNDLE' && preparedEvent.body) {
-                    // Step 1: Normalize legacy field names → canonical format
-                    const normalizedBundle = normalizeReleaseBundle(preparedEvent.body);
-
-                    // Step 2: Validate against canonical schema
-                    validateReleaseBundleOrThrow(normalizedBundle);
-
-                    // Replace event body with normalized+validated version
-                    preparedEvent.body = normalizedBundle;
-                }
-
-                // Calculate canonical hash (using same logic as EventStore)
-                // Exclude sig field to match storage hash calculation
-                const hash = this.store.calculateHash(preparedEvent);
-
-                // CRITICAL: Get canonical payload for signing
-                // Frontend must sign this exact string, not the hash
-                // This ensures signature verification will succeed in backend
-                const canonical_payload = this.store.getCanonicalPayload(preparedEvent);
-
-                res.json({
-                    success: true,
-                    hash,
-                    normalizedEvent: preparedEvent,
-                    canonical_payload
-                });
-            } catch (error) {
-                console.error('Event preparation failed:', error);
-                res.status(400).json({
-                    success: false,
-                    error: error.message
-                });
-            }
-        });
-
-        /**
-         * GET /api/events/dev-pubkey
-         * Development-only endpoint for getting the dev signer public key
-         *
-         * WARNING: DEV/TEST ONLY!
-         * - Only enabled when DEV_SIGNER_PRIVATE_KEY is set
-         * - Disabled in production (NODE_ENV=production)
-         * - Returns 403 if not enabled
-         *
-         * This allows smoke tests to know the author_pubkey before calling
-         * /api/events/prepare, ensuring the canonical payload includes it.
-         *
-         * Response:
-         * {
-         *   "success": true,
-         *   "author_pubkey": "EOS..."
-         * }
-         */
-        this.app.get('/api/events/dev-pubkey', (req, res) => {
-            try {
-                const devSigner = getDevSigner();
-                if (!devSigner?.isEnabled?.() || !devSigner.getPublicKey?.()) {
-                    return res.status(403).json({
-                        success: false,
-                        error: 'DevSigner not enabled. Set DEV_SIGNER_PRIVATE_KEY and ensure NODE_ENV !== production'
-                    });
-                }
-
-                return res.json({
-                    success: true,
-                    author_pubkey: devSigner.getPublicKey()
-                });
-            } catch (error) {
-                return res.status(500).json(sanitizeError(error, req.requestId, { success: false, env: this.config.env }));
-            }
-        });
-
-        /**
-         * POST /api/events/dev-sign
-         * Development-only endpoint for signing event canonical payloads
-         *
-         * WARNING: DEV/TEST ONLY!
-         * - Only enabled when DEV_SIGNER_PRIVATE_KEY is set
-         * - Disabled in production (NODE_ENV=production)
-         * - Returns 404 if not enabled
-         *
-         * This allows smoke tests and development workflows to create properly
-         * signed events without requiring a wallet/frontend integration.
-         *
-         * In production, signing must happen client-side with user's private keys.
-         *
-         * Request body:
-         * {
-         *   "canonical_payload": "string" - Canonical JSON string from /api/events/prepare
-         * }
-         *
-         * Response:
-         * {
-         *   "success": true,
-         *   "sig": "SIG_K1_...",
-         *   "author_pubkey": "EOS..."
-         * }
-         */
-        this.app.post('/api/events/dev-sign', async (req, res) => {
-            const devSigner = getDevSigner();
-
-            // Return 404 if dev signer not enabled (so it's not discoverable in prod)
-            if (!devSigner.isEnabled()) {
-                return res.status(404).json({
-                    success: false,
-                    error: 'Not found'
-                });
-            }
-
-            try {
-                const { canonical_payload } = req.body;
-
-                // Validate input
-                if (!canonical_payload || typeof canonical_payload !== 'string') {
-                    return res.status(400).json({
-                        success: false,
-                        error: 'canonical_payload is required and must be a string'
-                    });
-                }
-
-                if (canonical_payload.length === 0) {
-                    return res.status(400).json({
-                        success: false,
-                        error: 'canonical_payload cannot be empty'
-                    });
-                }
-
-                // Sign the canonical payload
-                const { sig, author_pubkey } = devSigner.signCanonicalPayload(canonical_payload);
-
-                res.json({
-                    success: true,
-                    sig,
-                    author_pubkey
-                });
-            } catch (error) {
-                console.error('Dev signing failed:', error);
-                res.status(500).json(sanitizeError(error, req.requestId, { success: false, env: this.config.env }));
-            }
-        });
-
-        /**
-         * POST /api/crypto/resolve-signing-key
-         * Determine which of an account's permission keys produced a given signature.
-         *
-         * When the wallet does not return the signing public key (signingKey is null
-         * on the frontend), the frontend calls this endpoint so the backend can
-         * try each key in the permission and return the one that verifies.
-         *
-         * Request body:
-         * {
-         *   "account":           "alice",            // blockchain account name
-         *   "permission":        "active",           // permission to inspect
-         *   "canonical_payload": "<json-string>",    // canonical JSON that was signed
-         *   "signature":         "SIG_K1_..."        // EOSIO signature to verify
-         * }
-         *
-         * Response (200):
-         * { "success": true, "signing_key": "EOS6..." }
-         *
-         * Response (400): missing/invalid fields
-         * Response (404): no matching key found
-         */
-        this.app.post('/api/crypto/resolve-signing-key', async (req, res) => {
-            try {
-                // RPC is required for key resolution — fail clearly if unavailable
-                const rpcUrl = process.env.RPC_URL || this.config.rpcUrl;
-                if (!rpcUrl) {
-                    return res.status(503).json({
-                        success: false,
-                        error: 'RPC_URL not configured — cannot resolve signing keys'
-                    });
-                }
-
-                const { account, permission, canonical_payload, signature } = req.body;
-
-                // Validate required fields
-                if (!account || !permission || !canonical_payload || !signature) {
-                    return res.status(400).json({
-                        success: false,
-                        error: 'Missing required fields: account, permission, canonical_payload, signature'
-                    });
-                }
-
-                // Hash the canonical payload (same as verifyEventSignature)
-                const payloadHash = createHash('sha256')
-                    .update(canonical_payload)
-                    .digest();
-
-                // Parse the signature once
-                let sig;
-                try {
-                    sig = Signature.fromString(signature);
-                } catch (parseErr) {
-                    return res.status(400).json({
-                        success: false,
-                        error: `Invalid signature format: ${parseErr.message}`
-                    });
-                }
-
-                // Fetch account permissions via IngestionHandler's cached RPC helper
-                const accountData = await this.ingestionHandler.fetchAccountData(account);
-                if (!accountData) {
-                    return res.status(400).json({
-                        success: false,
-                        error: `Cannot fetch account data for '${account}'`
-                    });
-                }
-
-                const perm = accountData.permissions?.find(p => p.perm_name === permission);
-                if (!perm) {
-                    return res.status(400).json({
-                        success: false,
-                        error: `Permission '${permission}' not found for account '${account}'`
-                    });
-                }
-
-                const keys = perm.required_auth?.keys || [];
-                if (keys.length === 0) {
-                    return res.status(404).json({
-                        success: false,
-                        error: `No keys found in permission '${permission}' for account '${account}'`
-                    });
-                }
-
-                // Try each key until one verifies
-                for (const keyEntry of keys) {
-                    try {
-                        const pubKey = PublicKey.fromString(keyEntry.key);
-                        if (sig.verify(payloadHash, pubKey)) {
-                            return res.json({
-                                success: true,
-                                signing_key: keyEntry.key
-                            });
-                        }
-                    } catch {
-                        // Key parse/verify failure — skip to next key
-                    }
-                }
-
-                // None matched
-                return res.status(404).json({
-                    success: false,
-                    error: 'Signature does not match any key in the specified permission'
-                });
-            } catch (error) {
-                console.error('resolve-signing-key failed:', error);
-                res.status(500).json(sanitizeError(error, req.requestId, { success: false, env: this.config.env }));
-            }
-        });
-
-        /**
-         * POST /api/events/create
-         * Submit a new event to storage and blockchain
-         *
-         * Validates CREATE_RELEASE_BUNDLE events against canonical schema
-         * to ensure no partial writes and deterministic error messages.
-         *
-         * @param {string} req.body.expected_hash - Optional. Expected hash from /api/events/prepare.
-         *                                          If provided and doesn't match computed hash, returns 400.
-         */
-        this.app.post('/api/events/create', this.writeRateLimiter, async (req, res) => {
-            const evtLog = createLogger('api.events.create', { request_id: req.requestId });
-            const timer = evtLog.startTimer();
-            try {
-                const { expected_hash, ...event } = req.body;
-                evtLog.info('event_create_start', { event_type: event.type, expected_hash: expected_hash || undefined });
-
-                // Validate CREATE_RELEASE_BUNDLE events at ingress
-                if (event.type === 'CREATE_RELEASE_BUNDLE' && event.body) {
-                    // Step 1: Normalize legacy field names → canonical format
-                    const normalizedBundle = normalizeReleaseBundle(event.body);
-
-                    // Step 2: Validate against canonical schema
-                    // This ensures deterministic validation and prevents partial writes
-                    validateReleaseBundleOrThrow(normalizedBundle);
-
-                    // Replace event body with normalized+validated version
-                    event.body = normalizedBundle;
-                }
-
-                // Store validated event (pass expectedHash for verification)
-                const result = await this.store.storeEvent(event, expected_hash || null);
-
-                // CRITICAL: event_cid is REQUIRED for blockchain anchoring
-                // If storage succeeded but event_cid is missing, return 503 (Service Unavailable)
-                // This prevents the frontend from attempting blockchain submission with null event_cid
-                if (!result.event_cid) {
-                    console.error('Event storage incomplete: missing event_cid');
-                    return res.status(503).json({
-                        success: false,
-                        error: 'IPFS required: could not produce event_cid for blockchain anchoring',
-                        hash: result.hash,
-                        stored: {
-                            canonical_cid: result.canonical_cid,
-                            event_cid: result.event_cid,
-                            s3: result.s3,
-                            redis: result.redis
-                        },
-                        replication: result.replication || { canonical: {}, event: {} },
-                        pinning: result.pinning || { attempted: false, success: false },
-                        errors: result.errors,
-                        message: 'Event stored to fallback storage but cannot be anchored on blockchain without IPFS event_cid. Check IPFS daemon status.'
-                    });
-                }
-
-                timer.end('event_create_end', { event_hash: result.hash, event_cid: result.event_cid });
-                res.status(201).json({
-                    success: true,
-                    hash: result.hash,
-                    stored: {
-                        canonical_cid: result.canonical_cid,
-                        event_cid: result.event_cid,
-                        s3: result.s3,
-                        redis: result.redis
-                    },
-                    replication: result.replication || { canonical: {}, event: {} },
-                    pinning: result.pinning || { attempted: false, success: false },
-                    errors: result.errors
-                });
-            } catch (error) {
-                timer.endError('event_create_error', { error: error.message });
-                res.status(400).json({
-                    success: false,
-                    error: error.message
-                });
-            }
-        });
-
-        /**
-         * GET /api/events/:hash
-         * Retrieve an event by its hash
-         */
-        this.app.get('/api/events/:hash', async (req, res) => {
-            const retLog = createLogger('api.events.retrieve', { request_id: req.requestId });
-            const timer = retLog.startTimer();
-            try {
-                retLog.info('event_retrieve_start', { event_hash: req.params.hash });
-                const event = await this.store.retrieveEvent(req.params.hash);
-
-                timer.end('event_retrieve_end', { event_hash: req.params.hash });
-                res.json({
-                    success: true,
-                    event
-                });
-            } catch (error) {
-                timer.endError('event_retrieve_error', { event_hash: req.params.hash, error: error.message });
-                res.status(404).json({
-                    success: false,
-                    error: error.message
-                });
-            }
-        });
-
-        // NOTE: /api/merge has been removed. Merges go through:
-        //   - Chain mode: POST /api/events/prepare → sign → POST /api/events/create → anchor on-chain
-        //   - Dev mode: POST /api/identity/merge (applies immediately, stores event for replay)
-        // See docs/12-identity-protocol.md for the canonical merge workflow.
-
-        // ========== ANCHOR-AUTH STORAGE ENDPOINT ==========
-
-        /**
-         * POST /api/events/store-for-anchor
-         * Store an event using anchor-auth flow (no off-chain signature required).
-         *
-         * In this flow, the on-chain put() transaction serves as the authoritative
-         * proof of authorship, replacing the unsupported WharfKit signMessage() call.
-         *
-         * The event is stored without `sig` and linked to a blockchain transaction
-         * via the subsequent POST /api/events/confirm-anchor call.
-         *
-         * Request body: Event object (without sig) + optional expected_hash
-         * Response: Same as /api/events/create
-         */
-        this.app.post('/api/events/store-for-anchor', this.writeRateLimiter, async (req, res) => {
-            const evtLog = createLogger('api.events.storeForAnchor', { request_id: req.requestId });
-            const timer = evtLog.startTimer();
-            try {
-                const { expected_hash, author_account, author_permission, ...event } = req.body;
-                evtLog.info('store_for_anchor_start', {
-                    event_type: event.type,
-                    author_account,
-                    expected_hash: expected_hash || undefined
-                });
-
-                if (!author_account) {
-                    return res.status(400).json({
-                        success: false,
-                        error: 'author_account is required for anchor-auth flow'
-                    });
-                }
-
-                // Validate CREATE_RELEASE_BUNDLE events at ingress
-                if (event.type === 'CREATE_RELEASE_BUNDLE' && event.body) {
-                    const normalizedBundle = normalizeReleaseBundle(event.body);
-                    validateReleaseBundleOrThrow(normalizedBundle);
-                    event.body = normalizedBundle;
-                }
-
-                // Store event without sig (anchor-auth mode)
-                const result = await this.store.storeEvent(event, expected_hash || null, { anchorAuth: true });
-
-                if (!result.event_cid) {
-                    return res.status(503).json({
-                        success: false,
-                        error: 'IPFS required: could not produce event_cid for blockchain anchoring',
-                        hash: result.hash,
-                        stored: {
-                            canonical_cid: result.canonical_cid,
-                            event_cid: result.event_cid,
-                            s3: result.s3,
-                            redis: result.redis
-                        }
-                    });
-                }
-
-                timer.end('store_for_anchor_end', { event_hash: result.hash, event_cid: result.event_cid });
-                res.status(201).json({
-                    success: true,
-                    hash: result.hash,
-                    stored: {
-                        canonical_cid: result.canonical_cid,
-                        event_cid: result.event_cid,
-                        s3: result.s3,
-                        redis: result.redis
-                    },
-                    replication: result.replication || { canonical: {}, event: {} },
-                    pinning: result.pinning || { attempted: false, success: false },
-                    errors: result.errors
-                });
-            } catch (error) {
-                timer.endError('store_for_anchor_error', { error: error.message });
-                res.status(400).json({
-                    success: false,
-                    error: error.message
-                });
-            }
-        });
-
-        /**
-         * POST /api/events/confirm-anchor
-         * Confirm that a stored event has been anchored on-chain via put() transaction.
-         *
-         * This endpoint links the stored event to its blockchain transaction,
-         * completing the anchor-auth flow. The on-chain transaction serves as
-         * the authoritative proof of authorship.
-         *
-         * Request body:
-         * {
-         *   hash: string,            // Event hash
-         *   event_cid: string,       // IPFS CID of event
-         *   trx_id: string,          // Blockchain transaction ID
-         *   author_account: string,  // Blockchain account that submitted the tx
-         *   author_permission: string // Permission used (e.g. "active")
-         * }
-         */
-        this.app.post('/api/events/confirm-anchor', this.writeRateLimiter, async (req, res) => {
-            const evtLog = createLogger('api.events.confirmAnchor', { request_id: req.requestId });
-            try {
-                const { hash, event_cid, trx_id, author_account, author_permission } = req.body;
-
-                if (!hash || !trx_id || !author_account) {
-                    return res.status(400).json({
-                        success: false,
-                        error: 'hash, trx_id, and author_account are required'
-                    });
-                }
-
-                evtLog.info('confirm_anchor', {
-                    event_hash: hash,
-                    trx_id,
-                    author_account,
-                    author_permission: author_permission || 'active',
-                    event_cid: event_cid || ''
-                });
-
-                // Store anchor confirmation metadata in Redis if available
-                if (this.store.redis) {
-                    try {
-                        const anchorKey = `anchor:${hash}`;
-                        await this.store.redis.hmset(anchorKey, {
-                            trx_id,
-                            author_account,
-                            author_permission: author_permission || 'active',
-                            event_cid: event_cid || '',
-                            confirmed_at: Math.floor(Date.now() / 1000)
-                        });
-                        await this.store.redis.expire(anchorKey, 30 * 24 * 60 * 60); // 30 days TTL
-                    } catch (redisErr) {
-                        evtLog.warn('confirm_anchor_redis_error', { error: redisErr.message });
-                    }
-                }
-
-                res.json({
-                    success: true,
-                    hash,
-                    trx_id,
-                    author_account,
-                    status: 'anchored'
-                });
-            } catch (error) {
-                evtLog.error('confirm_anchor_error', { error: error.message });
-                res.status(500).json(sanitizeError(error, req.requestId, { success: false, env: this.config.env }));
-            }
-        });
-
-        // ========== CHAIN READER ENDPOINTS ==========
-
-        // Register chain reader routes (likes, vote tallies)
-        // These proxy blockchain table reads through the backend so the browser
-        // never needs direct RPC access to chain nodes (eliminates CSP issues)
+        // Chain reader (likes, vote tallies). Registers its own routes on app.
         this.chainReader.registerRoutes(this.app);
 
-        // ========== CHAIN INGESTION ENDPOINT (T5) ==========
+        // Chain ingestion endpoint
+        this.app.use('/api/ingest', createIngestRoutes(ctx));
 
-        /**
-         * POST /api/ingest/anchored-event
-         * Ingest anchored event from Substreams chain ingestion
-         *
-         * Request body: AnchoredEvent
-         * {
-         *   content_hash: string,        // REQUIRED: Canonical content hash from put.hash
-         *   payload: string | Buffer,    // REQUIRED: Raw action JSON payload
-         *   event_hash: string,          // OPTIONAL: Action payload hash (debugging only)
-         *   block_num: number,
-         *   block_id: string,
-         *   trx_id: string,
-         *   action_ordinal: number,
-         *   timestamp: number,
-         *   source: string,
-         *   contract_account: string,
-         *   action_name: string
-         * }
-         *
-         * Response:
-         * {
-         *   status: "processed" | "duplicate" | "error",
-         *   eventHash: string,
-         *   eventType?: string,
-         *   blockNum?: number,
-         *   trxId?: string,
-         *   processing?: Object
-         * }
-         */
-        this.app.post('/api/ingest/anchored-event', this.writeRateLimiter, this.requireApiKey, async (req, res) => {
-            const ingestLog = createLogger('api.ingest', { request_id: req.requestId });
-            const timer = ingestLog.startTimer();
-            try {
-                const anchoredEvent = req.body.anchoredEvent || req.body;
+        // Entity-detail endpoints (group, person, track, release, song, label).
+        // Each route inside the router carries its own /<entity>/:id prefix,
+        // so we mount the router at /api.
+        this.app.use('/api', createEntityRoutes(ctx));
 
-                // Validate required fields (content_hash is canonical, event_hash is optional)
-                if (!anchoredEvent.content_hash || !anchoredEvent.payload) {
-                    return res.status(400).json({
-                        status: 'error',
-                        error: 'Missing required fields: content_hash and payload'
-                    });
-                }
+        // Node search
+        this.app.use('/api/search', createSearchRoutes(ctx));
 
-                ingestLog.info('ingest_start', {
-                    event_hash: anchoredEvent.content_hash,
-                    source: anchoredEvent.source,
-                    block_num: anchoredEvent.block_num,
-                    trx_id: anchoredEvent.trx_id,
-                    action_name: anchoredEvent.action_name
-                });
+        // Curation
+        this.app.use('/api/curate', createCurateRoutes(ctx));
 
-                // Process anchored event
-                const result = await this.ingestionHandler.processAnchoredEvent(anchoredEvent);
+        // Player queue
+        this.app.use('/api/player', createPlayerRoutes(ctx));
 
-                // Return appropriate status code
-                const statusCode = result.status === 'duplicate' ? 200 : 201;
-
-                timer.end('ingest_end', {
-                    event_hash: anchoredEvent.content_hash,
-                    result_status: result.status,
-                    status: statusCode
-                });
-
-                res.status(statusCode).json(result);
-
-            } catch (error) {
-                timer.endError('ingest_error', {
-                    error: error.message,
-                    error_class: error.constructor.name
-                });
-                res.status(500).json(sanitizeError(error, req.requestId, { env: this.config.env }));
-            }
-        });
-
-        // ========== GROUP ENDPOINTS ==========
-
-        /**
-         * GET /api/groups/:groupId/participation
-         * Get member participation data for RGraph visualization
-         */
-        this.app.get('/api/groups/:groupId/participation', async (req, res) => {
-            const groupId = req.params.groupId;
-            const startMs = Date.now();
-            try {
-                const rows = await this.db.calculateGroupMemberParticipation(groupId);
-
-                // rows is a flat array; totalTracks is the same on every row
-                const totalTracks = rows.length > 0 ? rows[0].totalTracks : 0;
-                const members = rows.map(r => ({
-                    personId: r.personId,
-                    personName: r.personName,
-                    trackCount: r.trackCount,
-                    trackPctOfGroupTracks: r.trackPctOfGroupTracks,
-                    color: r.color ?? null
-                }));
-
-                const durationMs = Date.now() - startMs;
-                console.log(`Participation request: group=${groupId.substring(0, 12)} members=${members.length} duration=${durationMs}ms`);
-
-                res.json({
-                    success: true,
-                    groupId,
-                    totalTracks,
-                    members
-                });
-            } catch (error) {
-                const durationMs = Date.now() - startMs;
-                console.error(`Participation request failed: group=${groupId.substring(0, 12)} duration=${durationMs}ms`, error);
-                res.status(500).json(sanitizeError(error, req.requestId, { success: false, env: this.config.env }));
-            }
-        });
-
-        /**
-         * GET /api/person/:personId
-         * Get person details
-         */
-        this.app.get('/api/person/:personId', async (req, res) => {
-            try {
-                const session = this.db.driver.session();
-                try {
-                    const result = await session.run(`
-                        MATCH (p:Person {person_id: $personId})
-                        OPTIONAL MATCH (p)-[m:MEMBER_OF]->(g:Group)
-                        OPTIONAL MATCH (p)-[:WROTE]->(s:Song)
-                        OPTIONAL MATCH (p)-[:GUEST_ON]->(t:Track)
-
-                        RETURN p,
-                               collect(DISTINCT {
-                                   group: g.name,
-                                   group_id: g.group_id,
-                                   role: m.role,
-                                   from_date: m.from_date,
-                                   to_date: m.to_date
-                               }) as groups,
-                               count(DISTINCT s) as songsWritten,
-                               count(DISTINCT t) as guestAppearances
-                    `, { personId: req.params.personId });
-
-                    if (result.records.length === 0) {
-                        return res.status(404).json({
-                            success: false,
-                            error: 'Person not found'
-                        });
-                    }
-
-                    const record = result.records[0];
-                    const person = record.get('p').properties;
-
-                    res.json({
-                        success: true,
-                        data: {
-                            ...person,
-                            groups: record.get('groups').filter(g => g.group !== null),
-                            songsWritten: record.get('songsWritten').toNumber(),
-                            guestAppearances: record.get('guestAppearances').toNumber()
-                        }
-                    });
-                } finally {
-                    await session.close();
-                }
-            } catch (error) {
-                console.error('Person details failed:', error);
-                res.status(500).json(sanitizeError(error, req.requestId, { success: false, env: this.config.env }));
-            }
-        });
-
-        /**
-         * GET /api/group/:groupId
-         * Get group details
-         */
-        this.app.get('/api/group/:groupId', async (req, res) => {
-            try {
-                const session = this.db.driver.session();
-                try {
-                    const result = await session.run(`
-                        MATCH (g:Group {group_id: $groupId})
-                        OPTIONAL MATCH (g)-[:PERFORMED_ON]->(t:Track)
-                        OPTIONAL MATCH (t)-[:IN_RELEASE]->(r:Release)
-                        OPTIONAL MATCH (p:Person)-[m:MEMBER_OF]->(g)
-
-                        RETURN g,
-                               count(DISTINCT t) as trackCount,
-                               count(DISTINCT r) as releaseCount,
-                               collect(DISTINCT {
-                                   person: p.name,
-                                   person_id: p.person_id,
-                                   role: m.role,
-                                   from_date: m.from_date,
-                                   to_date: m.to_date
-                               }) as members
-                    `, { groupId: req.params.groupId });
-
-                    if (result.records.length === 0) {
-                        return res.status(404).json({
-                            success: false,
-                            error: 'Group not found'
-                        });
-                    }
-
-                    const record = result.records[0];
-                    const group = record.get('g').properties;
-
-                    res.json({
-                        success: true,
-                        data: {
-                            ...group,
-                            trackCount: record.get('trackCount').toNumber(),
-                            releaseCount: record.get('releaseCount').toNumber(),
-                            members: record.get('members').filter(m => m.person !== null)
-                        }
-                    });
-                } finally {
-                    await session.close();
-                }
-            } catch (error) {
-                console.error('Group details failed:', error);
-                res.status(500).json(sanitizeError(error, req.requestId, { success: false, env: this.config.env }));
-            }
-        });
-
-        /**
-         * GET /api/group/:groupId/releases
-         * Get releases associated with a group (for release orbit overlay)
-         */
-        this.app.get('/api/group/:groupId/releases', async (req, res) => {
-            try {
-                const session = this.db.driver.session();
-                try {
-                    const result = await session.run(`
-                        MATCH (g:Group {group_id: $groupId})
-                        MATCH (g)-[:PERFORMED_ON]->(t:Track)-[:IN_RELEASE]->(r:Release)
-                        WITH g, r, count(DISTINCT t) as trackCount
-                        OPTIONAL MATCH (guest:Person)-[:GUEST_ON]->(gt:Track)-[:IN_RELEASE]->(r)
-                        WHERE NOT (guest)-[:MEMBER_OF]->(g)
-                        WITH r, trackCount, count(DISTINCT guest) as guestCount
-                        RETURN r.release_id as release_id,
-                               r.name as name,
-                               r.release_date as release_date,
-                               r.album_art as album_art,
-                               r.format as format,
-                               r.master_id as master_id,
-                               trackCount as track_count,
-                               guestCount as guest_count
-                        ORDER BY r.release_date ASC, r.name ASC
-                    `, { groupId: req.params.groupId });
-
-                    const releases = result.records.map(rec => ({
-                        release_id: rec.get('release_id'),
-                        name: rec.get('name'),
-                        release_date: rec.get('release_date'),
-                        album_art: rec.get('album_art'),
-                        format: rec.get('format'),
-                        master_id: rec.get('master_id'),
-                        track_count: typeof rec.get('track_count')?.toNumber === 'function'
-                            ? rec.get('track_count').toNumber() : rec.get('track_count'),
-                        guest_count: typeof rec.get('guest_count')?.toNumber === 'function'
-                            ? rec.get('guest_count').toNumber() : rec.get('guest_count')
-                    }));
-
-                    res.json({
-                        success: true,
-                        groupId: req.params.groupId,
-                        releases
-                    });
-                } finally {
-                    await session.close();
-                }
-            } catch (error) {
-                console.error('Group releases failed:', error);
-                res.status(500).json(sanitizeError(error, req.requestId, { success: false, env: this.config.env }));
-            }
-        });
-
-        /**
-         * GET /api/release/:releaseId
-         * Get release details (includes tracks, labels, groups, and guests)
-         */
-        this.app.get('/api/release/:releaseId', async (req, res) => {
-            try {
-                const session = this.db.driver.session();
-                try {
-                    // Main release + tracks + labels
-                    const result = await session.run(`
-                        MATCH (r:Release {release_id: $releaseId})
-                        OPTIONAL MATCH (t:Track)-[ir:IN_RELEASE]->(r)
-                        OPTIONAL MATCH (r)-[:RELEASED]->(l:Label)
-
-                        RETURN r,
-                               collect(DISTINCT {
-                                   track: t.title,
-                                   track_id: t.track_id,
-                                   disc_number: ir.disc_number,
-                                   track_number: ir.track_number,
-                                   side: ir.side
-                               }) as tracks,
-                               collect(DISTINCT {
-                                   label: l.name,
-                                   label_id: l.label_id
-                               }) as labels
-                    `, { releaseId: req.params.releaseId });
-
-                    if (result.records.length === 0) {
-                        return res.status(404).json({
-                            success: false,
-                            error: 'Release not found'
-                        });
-                    }
-
-                    const record = result.records[0];
-                    const release = record.get('r').properties;
-
-                    // Groups that performed on this release
-                    const groupResult = await session.run(`
-                        MATCH (t:Track)-[:IN_RELEASE]->(:Release {release_id: $releaseId})
-                        MATCH (g:Group)-[:PERFORMED_ON]->(t)
-                        RETURN DISTINCT g.group_id as group_id, g.name as name
-                    `, { releaseId: req.params.releaseId });
-
-                    const groups = groupResult.records.map(r => ({
-                        group_id: r.get('group_id'),
-                        name: r.get('name')
-                    }));
-
-                    // Guests: persons credited via GUEST_ON on tracks in this release
-                    const guestResult = await session.run(`
-                        MATCH (t:Track)-[:IN_RELEASE]->(:Release {release_id: $releaseId})
-                        MATCH (p:Person)-[go:GUEST_ON]->(t)
-                        WITH p, collect(DISTINCT coalesce(go.role, go.instrument, '')) as roles
-                        RETURN p.person_id as person_id,
-                               p.name as name,
-                               p.color as color,
-                               roles
-                    `, { releaseId: req.params.releaseId });
-
-                    const guests = guestResult.records.map(r => ({
-                        person_id: r.get('person_id'),
-                        name: r.get('name'),
-                        color: r.get('color'),
-                        roles: r.get('roles').filter(Boolean)
-                    }));
-
-                    res.json({
-                        success: true,
-                        data: {
-                            ...release,
-                            tracks: record.get('tracks').filter(t => t.track !== null),
-                            labels: record.get('labels').filter(l => l.label !== null),
-                            groups,
-                            guests
-                        }
-                    });
-                } finally {
-                    await session.close();
-                }
-            } catch (error) {
-                console.error('Release details failed:', error);
-                res.status(500).json(sanitizeError(error, req.requestId, { success: false, env: this.config.env }));
-            }
-        });
-
-        /**
-         * GET /api/track/:trackId
-         * Get track details
-         */
-        this.app.get('/api/track/:trackId', async (req, res) => {
-            try {
-                const session = this.db.driver.session();
-                try {
-                    const result = await session.run(`
-                        MATCH (t:Track {track_id: $trackId})
-                        OPTIONAL MATCH (t)-[:RECORDING_OF]->(s:Song)
-                        OPTIONAL MATCH (g:Group)-[:PERFORMED_ON]->(t)
-                        OPTIONAL MATCH (p:Person)-[:GUEST_ON]->(t)
-                        OPTIONAL MATCH (t)-[:IN_RELEASE]->(r:Release)
-
-                        RETURN t,
-                               s,
-                               collect(DISTINCT {
-                                   group: g.name,
-                                   group_id: g.group_id
-                               }) as performedBy,
-                               collect(DISTINCT {
-                                   guest: p.name,
-                                   person_id: p.person_id
-                               }) as guests,
-                               collect(DISTINCT {
-                                   release: r.name,
-                                   release_id: r.release_id
-                               }) as releases
-                    `, { trackId: req.params.trackId });
-
-                    if (result.records.length === 0) {
-                        return res.status(404).json({
-                            success: false,
-                            error: 'Track not found'
-                        });
-                    }
-
-                    const record = result.records[0];
-                    const track = record.get('t').properties;
-                    const song = record.get('s');
-
-                    res.json({
-                        success: true,
-                        data: {
-                            ...track,
-                            song: song ? song.properties : null,
-                            performedBy: record.get('performedBy').filter(g => g.group !== null),
-                            guests: record.get('guests').filter(g => g.guest !== null),
-                            releases: record.get('releases').filter(r => r.release !== null)
-                        }
-                    });
-                } finally {
-                    await session.close();
-                }
-            } catch (error) {
-                console.error('Track details failed:', error);
-                res.status(500).json(sanitizeError(error, req.requestId, { success: false, env: this.config.env }));
-            }
-        });
-
-        /**
-         * GET /api/song/:songId
-         * Get song details
-         */
-        this.app.get('/api/song/:songId', async (req, res) => {
-            try {
-                const session = this.db.driver.session();
-                try {
-                    const result = await session.run(`
-                        MATCH (s:Song {song_id: $songId})
-                        OPTIONAL MATCH (p:Person)-[:WROTE]->(s)
-                        OPTIONAL MATCH (t:Track)-[:RECORDING_OF]->(s)
-                        OPTIONAL MATCH (t)-[:IN_RELEASE]->(r:Release)
-
-                        RETURN s,
-                               collect(DISTINCT {
-                                   writer: p.name,
-                                   person_id: p.person_id
-                               }) as writers,
-                               collect(DISTINCT {
-                                   track: t.title,
-                                   track_id: t.track_id
-                               }) as recordings,
-                               collect(DISTINCT {
-                                   release: r.name,
-                                   release_id: r.release_id,
-                                   release_date: r.release_date,
-                                   album_art: r.album_art
-                               }) as releases
-                    `, { songId: req.params.songId });
-
-                    if (result.records.length === 0) {
-                        return res.status(404).json({
-                            success: false,
-                            error: 'Song not found'
-                        });
-                    }
-
-                    const record = result.records[0];
-                    const song = record.get('s').properties;
-
-                    res.json({
-                        success: true,
-                        data: {
-                            ...song,
-                            writers: record.get('writers').filter(w => w.writer !== null),
-                            recordings: record.get('recordings').filter(r => r.track !== null),
-                            releases: record.get('releases').filter(r => r.release !== null)
-                        }
-                    });
-                } finally {
-                    await session.close();
-                }
-            } catch (error) {
-                console.error('Song details failed:', error);
-                res.status(500).json(sanitizeError(error, req.requestId, { success: false, env: this.config.env }));
-            }
-        });
-
-        /**
-         * GET /api/label/:labelId
-         * Get label details
-         */
-        this.app.get('/api/label/:labelId', async (req, res) => {
-            try {
-                const session = this.db.driver.session();
-                try {
-                    const result = await session.run(`
-                        MATCH (l:Label {label_id: $labelId})
-                        OPTIONAL MATCH (l)<-[:RELEASED]-(r:Release)
-
-                        RETURN l,
-                               collect(DISTINCT {
-                                   release: r.name,
-                                   release_id: r.release_id,
-                                   release_date: r.release_date
-                               }) as releases,
-                               count(r) as releaseCount
-                    `, { labelId: req.params.labelId });
-
-                    if (result.records.length === 0) {
-                        return res.status(404).json({
-                            success: false,
-                            error: 'Label not found'
-                        });
-                    }
-
-                    const record = result.records[0];
-                    const label = record.get('l').properties;
-
-                    res.json({
-                        success: true,
-                        data: {
-                            ...label,
-                            releases: record.get('releases').filter(r => r.release !== null),
-                            releaseCount: record.get('releaseCount').toNumber()
-                        }
-                    });
-                } finally {
-                    await session.close();
-                }
-            } catch (error) {
-                console.error('Label details failed:', error);
-                res.status(500).json(sanitizeError(error, req.requestId, { success: false, env: this.config.env }));
-            }
-        });
-
-        /**
-         * GET /api/groups/:groupId/details
-         * Get comprehensive group information
-         */
-        this.app.get('/api/groups/:groupId/details', async (req, res) => {
-            try {
-                const session = this.db.driver.session();
-                try {
-                    const result = await session.run(`
-                        MATCH (g:Group {group_id: $groupId})
-                        OPTIONAL MATCH (g)-[:PERFORMED_ON]->(t:Track)
-                        OPTIONAL MATCH (t)-[:IN_RELEASE]->(r:Release)
-                        OPTIONAL MATCH (p:Person)-[m:MEMBER_OF]->(g)
-
-                        RETURN g,
-                               count(DISTINCT t) as trackCount,
-                               count(DISTINCT r) as releaseCount,
-                               collect(DISTINCT {
-                                   person: p.name,
-                                   role: m.role,
-                                   from: m.from_date,
-                                   to: m.to_date
-                               }) as members
-                    `, { groupId: req.params.groupId });
-
-                    if (result.records.length === 0) {
-                        return res.status(404).json({
-                            success: false,
-                            error: 'Group not found'
-                        });
-                    }
-
-                    const record = result.records[0];
-                    const group = record.get('g').properties;
-
-                    res.json({
-                        success: true,
-                        group: {
-                            ...group,
-                            trackCount: record.get('trackCount').toNumber(),
-                            releaseCount: record.get('releaseCount').toNumber(),
-                            members: record.get('members')
-                        }
-                    });
-                } finally {
-                    await session.close();
-                }
-            } catch (error) {
-                console.error('Group details failed:', error);
-                res.status(500).json(sanitizeError(error, req.requestId, { success: false, env: this.config.env }));
-            }
-        });
-
-        // ========== STATS ENDPOINTS ==========
-
-        /**
-         * GET /api/stats
-         * Get overall system statistics
-         */
-        this.app.get('/api/stats', async (req, res) => {
-            try {
-                const dbStats = await this.db.getStats();
-                const storageStats = this.store.getStats();
-
-                res.json({
-                    success: true,
-                    database: dbStats,
-                    storage: storageStats,
-                    timestamp: new Date().toISOString()
-                });
-            } catch (error) {
-                console.error('Stats retrieval failed:', error);
-                res.status(500).json(sanitizeError(error, req.requestId, { success: false, env: this.config.env }));
-            }
-        });
-
-        // ========== SEARCH ENDPOINT ==========
-
-        /**
-         * GET /api/search/nodes?q=<query>&types=Person,Group&limit=20
-         * Unified node search across user-facing entity labels.
-         */
-        this.app.get('/api/search/nodes', async (req, res) => {
-            try {
-                const q = req.query.q;
-                if (!q || q.trim().length < 2) {
-                    return res.json({ success: true, results: [] });
-                }
-
-                const types = req.query.types
-                    ? req.query.types.split(',').map(t => t.trim()).filter(Boolean)
-                    : [];
-                const limit = Math.min(parseInt(req.query.limit) || 20, 100);
-
-                const searchService = new NodeSearchService(this.db.driver);
-                const results = await searchService.search(q, { types, limit });
-
-                res.json({ success: true, results });
-            } catch (error) {
-                console.error('Node search failed:', error);
-                res.status(500).json(sanitizeError(error, req.requestId, { success: false, env: this.config.env }));
-            }
-        });
-
-        // ========== CURATION ENDPOINTS ==========
-
-        /**
-         * GET /api/curate/operations
-         * Get recent anchored operations with on-chain vote tallies.
-         * Reads anchors + votetally tables from the blockchain.
-         *
-         * Query params:
-         *   limit (default 50)
-         *   lower_bound (anchor id for pagination)
-         *   type (filter by event type)
-         */
-        this.app.get('/api/curate/operations', async (req, res) => {
-            try {
-                const rpcUrl = process.env.RPC_URL || this.config.rpcUrl;
-                if (!rpcUrl) {
-                    return res.status(503).json({ success: false, error: 'RPC_URL not configured' });
-                }
-
-                const contractAccount = process.env.CONTRACT_ACCOUNT || 'polaris';
-                const limit = Math.min(parseInt(req.query.limit) || 50, 200);
-                const lower_bound = req.query.lower_bound || undefined;
-
-                // Fetch anchors
-                const anchorsBody = {
-                    json: true,
-                    code: contractAccount,
-                    scope: contractAccount,
-                    table: 'anchors',
-                    limit,
-                    reverse: true // newest first
-                };
-                if (lower_bound !== undefined) anchorsBody.lower_bound = lower_bound;
-
-                const anchorsResp = await fetch(`${rpcUrl}/v1/chain/get_table_rows`, {
-                    method: 'POST',
-                    headers: { 'Content-Type': 'application/json' },
-                    body: JSON.stringify(anchorsBody)
-                });
-                if (!anchorsResp.ok) throw new Error(`Chain RPC error: ${anchorsResp.status}`);
-                const anchorsData = await anchorsResp.json();
-
-                // Filter by type if requested
-                let rows = anchorsData.rows || [];
-                if (req.query.type) {
-                    const filterType = parseInt(req.query.type);
-                    rows = rows.filter(r => r.type === filterType);
-                }
-
-                // Fetch tallies for each anchor
-                const operations = [];
-                for (const anchor of rows) {
-                    const tallyBody = {
-                        json: true,
-                        code: contractAccount,
-                        scope: contractAccount,
-                        table: 'votetally',
-                        limit: 1,
-                        lower_bound: String(anchor.id),
-                        upper_bound: String(anchor.id)
-                    };
-
-                    let tally = null;
-                    try {
-                        const tallyResp = await fetch(`${rpcUrl}/v1/chain/get_table_rows`, {
-                            method: 'POST',
-                            headers: { 'Content-Type': 'application/json' },
-                            body: JSON.stringify(tallyBody)
-                        });
-                        if (tallyResp.ok) {
-                            const tallyData = await tallyResp.json();
-                            tally = tallyData.rows && tallyData.rows[0] || null;
-                        }
-                    } catch (e) {
-                        // Tally fetch failure is non-fatal
-                    }
-
-                    // Try to get stored event summary from our event store
-                    let eventSummary = null;
-                    try {
-                        const stored = await this.store.retrieveEvent(anchor.hash);
-                        if (stored && stored.body) {
-                            eventSummary = {
-                                type_name: stored.type || null,
-                                release_name: stored.body?.release?.name || null,
-                                group_name: stored.body?.groups?.[0]?.name || null
-                            };
-                        }
-                    } catch (e) {
-                        // Event retrieval is best-effort
-                    }
-
-                    operations.push({
-                        anchor_id: anchor.id,
-                        author: anchor.author,
-                        type: anchor.type,
-                        hash: anchor.hash,
-                        event_cid: anchor.event_cid,
-                        ts: anchor.ts,
-                        expires_at: anchor.expires_at,
-                        finalized: anchor.finalized ? true : false,
-                        tally: tally ? {
-                            up_weight: parseInt(tally.up_weight) || 0,
-                            down_weight: parseInt(tally.down_weight) || 0,
-                            up_voter_count: parseInt(tally.up_voter_count) || 0,
-                            down_voter_count: parseInt(tally.down_voter_count) || 0
-                        } : { up_weight: 0, down_weight: 0, up_voter_count: 0, down_voter_count: 0 },
-                        event_summary: eventSummary
-                    });
-                }
-
-                res.json({
-                    success: true,
-                    operations,
-                    more: anchorsData.more || false,
-                    next_key: anchorsData.next_key || null
-                });
-            } catch (error) {
-                console.error('Curate operations failed:', error);
-                res.status(500).json(sanitizeError(error, req.requestId, { success: false, env: this.config.env }));
-            }
-        });
-
-        /**
-         * GET /api/curate/operations/:hash
-         * Get full details for a single anchored operation.
-         * Returns anchor, tally, stored event payload, and individual votes.
-         */
-        this.app.get('/api/curate/operations/:hash', async (req, res) => {
-            try {
-                const rpcUrl = process.env.RPC_URL || this.config.rpcUrl;
-                if (!rpcUrl) {
-                    return res.status(503).json({ success: false, error: 'RPC_URL not configured' });
-                }
-
-                const contractAccount = process.env.CONTRACT_ACCOUNT || 'polaris';
-                const hash = req.params.hash;
-
-                // Fetch anchor by hash (secondary index)
-                const anchorsResp = await fetch(`${rpcUrl}/v1/chain/get_table_rows`, {
-                    method: 'POST',
-                    headers: { 'Content-Type': 'application/json' },
-                    body: JSON.stringify({
-                        json: true,
-                        code: contractAccount,
-                        scope: contractAccount,
-                        table: 'anchors',
-                        index_position: 2, // byhash
-                        key_type: 'sha256',
-                        lower_bound: hash,
-                        upper_bound: hash,
-                        limit: 1
-                    })
-                });
-                if (!anchorsResp.ok) throw new Error(`Chain RPC error: ${anchorsResp.status}`);
-                const anchorsData = await anchorsResp.json();
-
-                if (!anchorsData.rows || anchorsData.rows.length === 0) {
-                    return res.status(404).json({ success: false, error: 'Anchor not found' });
-                }
-
-                const anchor = anchorsData.rows[0];
-
-                // Fetch tally
-                let tally = null;
-                try {
-                    const tallyResp = await fetch(`${rpcUrl}/v1/chain/get_table_rows`, {
-                        method: 'POST',
-                        headers: { 'Content-Type': 'application/json' },
-                        body: JSON.stringify({
-                            json: true,
-                            code: contractAccount,
-                            scope: contractAccount,
-                            table: 'votetally',
-                            limit: 1,
-                            lower_bound: String(anchor.id),
-                            upper_bound: String(anchor.id)
-                        })
-                    });
-                    if (tallyResp.ok) {
-                        const tallyData = await tallyResp.json();
-                        tally = tallyData.rows?.[0] || null;
-                    }
-                } catch (e) { /* non-fatal */ }
-
-                // Fetch individual votes for this hash
-                let votes = [];
-                try {
-                    const votesResp = await fetch(`${rpcUrl}/v1/chain/get_table_rows`, {
-                        method: 'POST',
-                        headers: { 'Content-Type': 'application/json' },
-                        body: JSON.stringify({
-                            json: true,
-                            code: contractAccount,
-                            scope: contractAccount,
-                            table: 'votes',
-                            index_position: 3, // byhash
-                            key_type: 'sha256',
-                            lower_bound: hash,
-                            upper_bound: hash,
-                            limit: 200
-                        })
-                    });
-                    if (votesResp.ok) {
-                        const votesData = await votesResp.json();
-                        votes = votesData.rows || [];
-                    }
-                } catch (e) { /* non-fatal */ }
-
-                // Fetch stored event payload
-                let eventPayload = null;
-                try {
-                    eventPayload = await this.store.retrieveEvent(anchor.hash);
-                } catch (e) { /* non-fatal */ }
-
-                // Parse type-specific detail for rendering
-                const detail = this._parseOperationDetail(eventPayload);
-
-                // Determine operation metadata
-                const typeCode = eventPayload?.type || anchor.type || 0;
-                const typeNames = {
-                    21: 'CREATE_RELEASE_BUNDLE', 30: 'ADD_CLAIM', 31: 'EDIT_CLAIM',
-                    40: 'VOTE', 41: 'LIKE', 50: 'FINALIZE', 60: 'MERGE_NODE'
-                };
-
-                // Check viewer's vote if account provided
-                const viewerAccount = req.query.viewer;
-                let viewerVote = null;
-                if (viewerAccount && votes.length > 0) {
-                    const found = votes.find(v => v.voter === viewerAccount);
-                    if (found) {
-                        viewerVote = { val: found.val, weight: found.weight };
-                    }
-                }
-
-                res.json({
-                    success: true,
-                    operation: {
-                        hash: anchor.hash,
-                        anchor_id: anchor.id,
-                        type_code: typeCode,
-                        type_name: typeNames[typeCode] || `TYPE_${typeCode}`,
-                        author: anchor.author || eventPayload?.author || null,
-                        ts: anchor.ts || null,
-                        finalized: !!anchor.finalized,
-                        event_cid: anchor.event_cid || null
-                    },
-                    tally: tally ? {
-                        up_weight: parseInt(tally.up_weight) || 0,
-                        down_weight: parseInt(tally.down_weight) || 0,
-                        up_voter_count: parseInt(tally.up_voter_count) || 0,
-                        down_voter_count: parseInt(tally.down_voter_count) || 0,
-                        updated_at: tally.updated_at
-                    } : { up_weight: 0, down_weight: 0, up_voter_count: 0, down_voter_count: 0 },
-                    viewer_vote: viewerVote,
-                    votes: votes.map(v => ({
-                        voter: v.voter,
-                        val: v.val,
-                        weight: v.weight,
-                        ts: v.ts
-                    })),
-                    event: eventPayload,
-                    detail
-                });
-            } catch (error) {
-                console.error('Curate operation detail failed:', error);
-                res.status(500).json(sanitizeError(error, req.requestId, { success: false, env: this.config.env }));
-            }
-        });
-
-        // ========== PLAYER ENDPOINTS ==========
-
-        /**
-         * GET /api/player/queue
-         * Build a playback queue for a release, group, or person context.
-         *
-         * Query params:
-         *   contextType - 'release', 'group', or 'person'
-         *   contextId   - Entity ID (e.g. 'prov:release:...')
-         */
-        this.app.get('/api/player/queue', async (req, res) => {
-            try {
-                const { contextType, contextId } = req.query;
-                if (!contextType || !contextId) {
-                    return res.status(400).json({
-                        success: false,
-                        error: 'contextType and contextId are required'
-                    });
-                }
-
-                if (!['release', 'group', 'person'].includes(contextType)) {
-                    return res.status(400).json({
-                        success: false,
-                        error: 'contextType must be release, group, or person'
-                    });
-                }
-
-                const { context, queue } = await this.playerService.buildQueue(contextType, contextId);
-                res.json({ success: true, context, queue });
-            } catch (error) {
-                console.error('Player queue failed:', error);
-                res.status(500).json(sanitizeError(error, req.requestId, { success: false, env: this.config.env }));
-            }
-        });
-
-        // ========== GRAPH ENDPOINTS ==========
-
-        /**
-         * GET /api/graph/initial
-         * Get initial graph data for visualization using all groups with tracks
-         * and their member relationships. Includes per-group participation data
-         * so the frontend does not need separate per-group fetches.
-         */
-        this.app.get('/api/graph/initial', async (req, res) => {
-            try {
-                const session = this.db.driver.session();
-                try {
-                    // Return all groups that have performed on at least one track,
-                    // along with their MEMBER_OF person relationships and
-                    // per-member participation counts (for donut visualization).
-                    // No LIMIT — the full graph is needed for connected visualization.
-                    const result = await session.run(`
-                        MATCH (g:Group)-[:PERFORMED_ON]->(t:Track)
-                        WITH g, count(DISTINCT t) as trackCount
-                        ORDER BY trackCount DESC
-                        OPTIONAL MATCH (p:Person)-[m:MEMBER_OF]->(g)
-                        OPTIONAL MATCH (p)-[:PERFORMED_ON {via_group_id: g.group_id}]->(pt:Track)
-                        WITH g, trackCount, p, m, count(DISTINCT pt) as memberTrackCount
-
-                        RETURN collect(DISTINCT {
-                            id: g.group_id,
-                            name: g.name,
-                            type: 'group',
-                            trackCount: trackCount,
-                            photo: g.photo
-                        }) as groups,
-                        collect(DISTINCT CASE WHEN p IS NOT NULL THEN {
-                            id: p.person_id,
-                            name: p.name,
-                            type: 'person',
-                            color: p.color
-                        } ELSE null END) as persons,
-                        collect(DISTINCT CASE WHEN m IS NOT NULL THEN {
-                            source: p.person_id,
-                            target: g.group_id,
-                            type: 'MEMBER_OF',
-                            role: m.role,
-                            from_date: m.from_date,
-                            to_date: m.to_date,
-                            instruments: m.instruments
-                        } ELSE null END) as edges,
-                        collect(DISTINCT CASE WHEN p IS NOT NULL THEN {
-                            groupId: g.group_id,
-                            personId: p.person_id,
-                            personName: p.name,
-                            color: p.color,
-                            trackCount: memberTrackCount,
-                            totalTracks: trackCount
-                        } ELSE null END) as participationRows
-                    `);
-
-                    if (result.records.length === 0) {
-                        return res.json({
-                            success: true,
-                            nodes: [],
-                            edges: [],
-                            participation: {}
-                        });
-                    }
-
-                    const groups = result.records[0].get('groups');
-                    // Filter out nulls from OPTIONAL MATCH results
-                    const persons = result.records[0].get('persons').filter(p => p !== null);
-                    const edges = result.records[0].get('edges').filter(e => e !== null);
-
-                    // Build participation map keyed by groupId
-                    const participationRows = result.records[0].get('participationRows').filter(r => r !== null);
-                    const participation = {};
-                    for (const row of participationRows) {
-                        const gid = row.groupId;
-                        if (!participation[gid]) {
-                            participation[gid] = { totalTracks: row.totalTracks, members: [] };
-                        }
-                        const totalTracks = participation[gid].totalTracks;
-                        const tc = typeof row.trackCount === 'object' && row.trackCount.toNumber
-                            ? row.trackCount.toNumber() : Number(row.trackCount) || 0;
-                        const tt = typeof totalTracks === 'object' && totalTracks.toNumber
-                            ? totalTracks.toNumber() : Number(totalTracks) || 0;
-                        participation[gid].totalTracks = tt;
-                        participation[gid].members.push({
-                            personId: row.personId,
-                            personName: row.personName,
-                            color: row.color || null,
-                            trackCount: tc,
-                            trackPctOfGroupTracks: tt > 0 ? (tc / tt) * 100.0 : 0.0
-                        });
-                    }
-
-                    res.json({
-                        success: true,
-                        nodes: [...groups, ...persons],
-                        edges: edges,
-                        participation: participation
-                    });
-                } finally {
-                    await session.close();
-                }
-            } catch (error) {
-                console.error('Initial graph failed:', error);
-                res.status(500).json(sanitizeError(error, req.requestId, { success: false, env: this.config.env }));
-            }
-        });
-
-        /**
-         * GET /api/graph/neighborhood/:nodeId
-         * Fetch a node's neighborhood subgraph for dynamic graph merging.
-         * Returns {nodes, edges} in the same shape as /graph/initial.
-         */
-        this.app.get('/api/graph/neighborhood/:nodeId', async (req, res) => {
-            const nodeId = req.params.nodeId;
-            const limitTracks = Math.min(parseInt(req.query.limitTracks) || 25, 100);
-            const limitGroups = Math.min(parseInt(req.query.limitGroups) || 25, 100);
-            const limitPersons = Math.min(parseInt(req.query.limitPersons) || 50, 200);
-
-            try {
-                const session = this.db.driver.session();
-                try {
-                    let nodes = [];
-                    let edges = [];
-
-                    if (nodeId.includes(':person:') || nodeId.startsWith('prov:person:')) {
-                        // Person neighborhood: person + all groups they're MEMBER_OF
-                        const result = await session.run(`
-                            MATCH (p:Person {person_id: $nodeId})
-                            OPTIONAL MATCH (p)-[m:MEMBER_OF]->(g:Group)
-                            RETURN p,
-                                   collect(DISTINCT {
-                                       group: g,
-                                       role: m.role,
-                                       from_date: m.from_date,
-                                       to_date: m.to_date,
-                                       instruments: m.instruments
-                                   }) as memberships
-                            LIMIT 1
-                        `, { nodeId });
-
-                        if (result.records.length > 0) {
-                            const record = result.records[0];
-                            const p = record.get('p');
-                            if (p) {
-                                nodes.push({
-                                    id: p.properties.person_id,
-                                    name: p.properties.name,
-                                    type: 'person',
-                                    color: p.properties.color || null
-                                });
-
-                                const memberships = record.get('memberships').filter(m => m.group !== null);
-                                for (const m of memberships.slice(0, limitGroups)) {
-                                    const g = m.group;
-                                    nodes.push({
-                                        id: g.properties.group_id,
-                                        name: g.properties.name,
-                                        type: 'group',
-                                        photo: g.properties.photo || null
-                                    });
-                                    edges.push({
-                                        source: p.properties.person_id,
-                                        target: g.properties.group_id,
-                                        type: 'MEMBER_OF',
-                                        role: m.role,
-                                        from_date: m.from_date,
-                                        to_date: m.to_date,
-                                        instruments: m.instruments
-                                    });
-                                }
-                            }
-                        }
-                    } else if (nodeId.includes(':group:') || nodeId.startsWith('prov:group:')) {
-                        // Group neighborhood: group + members + limited tracks + releases
-                        const result = await session.run(`
-                            MATCH (g:Group {group_id: $nodeId})
-                            OPTIONAL MATCH (p:Person)-[m:MEMBER_OF]->(g)
-                            WITH g, collect(DISTINCT {
-                                person: p,
-                                role: m.role,
-                                from_date: m.from_date,
-                                to_date: m.to_date,
-                                instruments: m.instruments
-                            }) as memberships
-                            OPTIONAL MATCH (g)-[:PERFORMED_ON]->(t:Track)
-                            WITH g, memberships, collect(DISTINCT t)[0..$limitTracks] as tracks
-                            UNWIND tracks as t
-                            OPTIONAL MATCH (t)-[:IN_RELEASE]->(r:Release)
-                            RETURN g, memberships,
-                                   collect(DISTINCT {track: t}) as trackNodes,
-                                   collect(DISTINCT {release: r, trackId: t.track_id}) as releaseEdges
-                        `, { nodeId, limitTracks: neo4j.int(limitTracks) });
-
-                        if (result.records.length > 0) {
-                            const record = result.records[0];
-                            const g = record.get('g');
-                            if (g) {
-                                nodes.push({
-                                    id: g.properties.group_id,
-                                    name: g.properties.name,
-                                    type: 'group',
-                                    photo: g.properties.photo || null
-                                });
-
-                                const memberships = record.get('memberships').filter(m => m.person !== null);
-                                for (const m of memberships.slice(0, limitPersons)) {
-                                    const p = m.person;
-                                    nodes.push({
-                                        id: p.properties.person_id,
-                                        name: p.properties.name,
-                                        type: 'person',
-                                        color: p.properties.color || null
-                                    });
-                                    edges.push({
-                                        source: p.properties.person_id,
-                                        target: g.properties.group_id,
-                                        type: 'MEMBER_OF',
-                                        role: m.role,
-                                        from_date: m.from_date,
-                                        to_date: m.to_date,
-                                        instruments: m.instruments
-                                    });
-                                }
-
-                                const trackNodes = record.get('trackNodes').filter(t => t.track !== null);
-                                for (const tn of trackNodes) {
-                                    const t = tn.track;
-                                    nodes.push({
-                                        id: t.properties.track_id,
-                                        name: t.properties.title || t.properties.name,
-                                        type: 'track'
-                                    });
-                                    edges.push({
-                                        source: g.properties.group_id,
-                                        target: t.properties.track_id,
-                                        type: 'PERFORMED_ON'
-                                    });
-                                }
-
-                                const releaseEdges = record.get('releaseEdges').filter(re => re.release !== null);
-                                const seenReleases = new Set();
-                                for (const re of releaseEdges) {
-                                    const r = re.release;
-                                    if (!seenReleases.has(r.properties.release_id)) {
-                                        seenReleases.add(r.properties.release_id);
-                                        nodes.push({
-                                            id: r.properties.release_id,
-                                            name: r.properties.name,
-                                            type: 'release'
-                                        });
-                                    }
-                                    edges.push({
-                                        source: re.trackId,
-                                        target: r.properties.release_id,
-                                        type: 'IN_RELEASE'
-                                    });
-                                }
-                            }
-                        }
-                    } else {
-                        // Generic fallback: try known node types
-                        const nodeTypes = [
-                            { label: 'Person', idField: 'person_id', type: 'person' },
-                            { label: 'Group', idField: 'group_id', type: 'group' },
-                            { label: 'Track', idField: 'track_id', type: 'track' },
-                            { label: 'Release', idField: 'release_id', type: 'release' },
-                            { label: 'Song', idField: 'song_id', type: 'song' }
-                        ];
-                        for (const meta of nodeTypes) {
-                            const result = await session.run(
-                                `MATCH (n:${meta.label} {${meta.idField}: $nodeId}) RETURN n LIMIT 1`,
-                                { nodeId }
-                            );
-                            if (result.records.length > 0) {
-                                const n = result.records[0].get('n');
-                                nodes.push({
-                                    id: n.properties[meta.idField],
-                                    name: n.properties.name || n.properties.title || 'Unknown',
-                                    type: meta.type
-                                });
-                                break;
-                            }
-                        }
-                    }
-
-                    // Deduplicate nodes by id
-                    const nodeMap = new Map();
-                    for (const n of nodes) {
-                        if (n && n.id) nodeMap.set(n.id, n);
-                    }
-
-                    res.json({
-                        success: true,
-                        nodes: Array.from(nodeMap.values()),
-                        edges
-                    });
-                } finally {
-                    await session.close();
-                }
-            } catch (error) {
-                console.error('Neighborhood fetch failed:', error);
-                res.status(500).json(sanitizeError(error, req.requestId, { success: false, env: this.config.env }));
-            }
-        });
+        // Graph data (initial + neighborhood)
+        this.app.use('/api/graph', createGraphRoutes(ctx));
 
         console.log(' REST endpoints configured');
     }
 
     /**
-     * Setup error handling middleware
-     */
-    setupErrorHandling() {
-        // 404 handler
-        this.app.use((req, res) => {
-            res.status(404).json({
-                success: false,
-                error: 'Endpoint not found',
-                path: req.path
-            });
-        });
-
-        // Global error handler
-        this.app.use((err, req, res, next) => {
-            console.error('Unhandled error:', err);
-
-            res.status(err.status || 500).json({
-                success: false,
-                error: err.message || 'Internal server error',
-                ...(process.env.NODE_ENV === 'development' && { stack: err.stack })
-            });
-        });
-    }
-
-    /**
-     * Start the server
+     * Start the server.
      *
      * @returns {Promise<void>}
      */
@@ -2666,8 +177,8 @@ class APIServer {
         }
         console.log(' Database connected');
 
-        // Initialize database schema (constraints, indexes)
-        // Controlled by GRAPH_INIT_SCHEMA env var (default: true in dev, configurable in prod)
+        // Initialize database schema (constraints, indexes).
+        // Controlled by GRAPH_INIT_SCHEMA env var (default: true).
         const shouldInitSchema = process.env.GRAPH_INIT_SCHEMA !== 'false';
         if (shouldInitSchema) {
             try {
@@ -2675,8 +186,8 @@ class APIServer {
                 console.log(' Database schema initialized');
             } catch (error) {
                 console.error(' Schema initialization failed:', error.message);
-                // Don't throw - allow server to start even if schema init fails
-                // This is safe because queries will still work, just without constraints
+                // Don't throw - allow server to start even if schema init fails.
+                // Queries will still work, just without constraints.
                 console.warn('  Continuing without schema initialization');
             }
         } else {
@@ -2769,121 +280,7 @@ class APIServer {
     }
 
     /**
-     * Parse type-specific detail from stored event payload for rendering.
-     * Returns a structured object the frontend can render without deep
-     * knowledge of event internals.
-     */
-    _parseOperationDetail(event) {
-        if (!event || !event.body) return null;
-
-        const typeCode = event.type;
-        const body = event.body;
-
-        if (typeCode === 21) {
-            // CREATE_RELEASE_BUNDLE
-            return {
-                type: 'release_bundle',
-                release: {
-                    name: body.release?.name || null,
-                    alt_names: body.release?.alt_names || [],
-                    release_date: body.release?.release_date || null,
-                    format: body.release?.format || null,
-                    liner_notes: body.release?.liner_notes || null,
-                    master_id: body.release?.master_id || null,
-                    album_art: body.release?.album_art || null,
-                    labels: (body.release?.labels || []).map(l => ({
-                        name: l.name,
-                        label_id: l.label_id || null,
-                        parent_label: l.parent_label || null
-                    })),
-                    guests: (body.release?.guests || []).map(g => ({
-                        name: g.name,
-                        person_id: g.person_id || null,
-                        roles: g.roles || []
-                    }))
-                },
-                groups: (body.groups || []).map(g => ({
-                    name: g.name,
-                    group_id: g.group_id || null,
-                    alt_names: g.alt_names || [],
-                    members: (g.members || []).map(m => ({
-                        name: m.name,
-                        person_id: m.person_id || null,
-                        roles: m.roles || []
-                    }))
-                })),
-                tracks: (body.tracks || []).map(t => ({
-                    title: t.title,
-                    track_id: t.track_id || null,
-                    recording_of: t.recording_of || null,
-                    song_id: t.song_id || null,
-                    listen_links: t.listen_links || [],
-                    cover_of_song_id: t.cover_of_song_id || null,
-                    samples: t.samples || [],
-                    performed_by_groups: (t.performed_by_groups || []).map(g => ({
-                        name: g.name,
-                        group_id: g.group_id || null,
-                        members: (g.members || []).map(m => ({
-                            name: m.name,
-                            person_id: m.person_id || null,
-                            roles: m.roles || []
-                        }))
-                    })),
-                    guests: (t.guests || []).map(g => ({
-                        name: g.name,
-                        person_id: g.person_id || null,
-                        roles: g.roles || []
-                    })),
-                    producers: (t.producers || []).map(p => ({
-                        name: p.name,
-                        person_id: p.person_id || null,
-                        roles: p.roles || []
-                    }))
-                })),
-                tracklist: body.tracklist || [],
-                songs: (body.songs || []).map(s => ({
-                    title: s.title,
-                    song_id: s.song_id || null,
-                    writers: (s.writers || []).map(w => ({
-                        name: w.name,
-                        person_id: w.person_id || null,
-                        roles: w.roles || []
-                    }))
-                })),
-                sources: body.sources || []
-            };
-        }
-
-        if (typeCode === 30) {
-            // ADD_CLAIM
-            return {
-                type: 'add_claim',
-                target_type: body.target_type || null,
-                target_id: body.target_id || null,
-                field: body.field || null,
-                value: body.value || null,
-                source: body.source || null
-            };
-        }
-
-        if (typeCode === 31) {
-            // EDIT_CLAIM
-            return {
-                type: 'edit_claim',
-                target_type: body.target_type || null,
-                target_id: body.target_id || null,
-                field: body.field || null,
-                value: body.value || null,
-                source: body.source || null
-            };
-        }
-
-        // Fallback for other types
-        return { type: `type_${typeCode}`, raw: body };
-    }
-
-    /**
-     * Stop the server and cleanup
+     * Stop the server and cleanup.
      *
      * @returns {Promise<void>}
      */
